+++ src/Writer/Strategy/Svg.php
@@ -25,12 +25,12 @@
 
 namespace PhpDA\Writer\Strategy;
 
-use PhpDA\Layout\GraphViz;
+use Fhaculty\Graph\Graph;
 
-class Svg implements StrategyInterface
+class Svg extends AbstractGraphViz
 {
-    public function filter(GraphViz $graphViz)
+    protected function toString(Graph $graph)
     {
-        return $graphViz->setFormat('svg')->createImageData();
+        return $this->getGraphViz()->setFormat('svg')->createImageData($graph);
     }
 }